extract GtinService into new microservice
