+++ services/gtinservice/.gitignore
@@ -1,0 +1,3 @@
+/target/
+/.settings/
+/.factorypath
