+++ services/gtinservice/README.md
@@ -1,0 +1,4 @@
+# GTIN Service
+
+This module provides utilities to generate EAN13 barcodes from GTIN identifiers.
+It exposes a `GtinService` bean used by other modules like the UI and static server.
