+++ static/pom.xml
@@ -61,6 +61,11 @@
     </dependency>
     <dependency>
       <groupId>org.open4goods</groupId>
+      <artifactId>gtinservice</artifactId>
+      <version>${global.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.open4goods</groupId>
       <artifactId>remotefilecaching</artifactId>
       <version>${global.version}</version>
     </dependency>
