+++ ui/pom.xml
@@ -133,6 +133,12 @@
             <version>${global.version}</version>
         </dependency>
 
+        <dependency>
+            <groupId>org.open4goods</groupId>
+            <artifactId>gtinservice</artifactId>
+            <version>${global.version}</version>
+        </dependency>
+
 
         <dependency>
             <groupId>org.open4goods</groupId>
